% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_taxonomy.R
\name{filter_taxonomy}
\alias{filter_taxonomy}
\title{filter taxonomy}
\usage{
filter_taxonomy(consensus, min_confidence = 0.8)
}
\arguments{
\item{consensus}{A list object that contains two slots each with
an equal sized vector. the \code{taxonomy} vector contains
the classification at each taxonomic level and the
\code{confidence} vector contains the fraction fo bootstraps
that the specified classification}

\item{min_confidence}{The minimum fraction of bootstrap replicates that had
the same classification. Any confidence score below this
value will have the corresponding taxonomy removed}
}
\value{
A list object containg two equally sized vectors that are filtered
to remove low confidence taxonomies. One vector, \code{taxonomy}, contains
the taxonomy at each taxonomic level and the other vector, \code{confidence}
contains the confidence score for that taxonomy.There will be no
taxonomies or confidence scores below \code{min_confidence}
}
\description{
The \code{filter_taxonomy()} function will filter a consensus taxonomy to remove
only taxonomic levels where the confidence score is below a \code{min_confidence}
level
}
\examples{
Oscillospiraceae <- list(taxonomy =c("Bacteria","Bacteroidota","Bacteroidia",
                                     "Bacteroidales","Porphyromonadaceae",
                                      "Macellibacteroides"),
                         confidence =c(1.00, 1.00, 1.00, 1.00, 0.83, 0.75))
filter_taxonomy(Oscillospiraceae,min_confidence = 0.80)

}
