% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_taxonomy.R
\name{print_taxonomy}
\alias{print_taxonomy}
\title{Print taxonomy for an unknown sequence}
\usage{
print_taxonomy(consensus, n_levels = 6)
}
\arguments{
\item{consensus}{A list object that contains two slots each with
an equal sized vector. the \code{taxonomy} vector contains
the classification at each taxonomic level and the
\code{confidence} vector contains the fraction fo bootstraps
that the specified classification}

\item{n_levels}{An integer that indicates the number of taxonomic levels to
expect for each sequence (default = 6)}
}
\value{
A character string indicating the classification at each taxonomic
level with the corresponding confidence in parentheses.
Each taxonomic level is separated by a semi-colon
}
\description{
The \code{print_taxonomy()} will output the cosensus taxonomy for an unknown
sequence will confidence scores for each taxonomic level and each
taxonomic level separated by semi-colons
}
\examples{
filtered <- list(taxonomy = c("Bacteria","Bacteroidota",
                 "Bacteroidia","Bacteroidales","Porphyromonadaceae"),
                  confidence = c(1.00,1.00,1.00,1.00,0.83))
tax_string <- print_taxonomy(filtered,n_levels = 6)

}
