% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_taxonomy.R
\name{read_taxonomy}
\alias{read_taxonomy}
\title{Read in taxonomy files}
\usage{
read_taxonomy(file)
}
\arguments{
\item{file}{Either a path to a file, a connection, or literal data
(either a single string or row vector) containing the sequence id
and the taxonomy information for each sequence.

files ending in .gz,.bz2, .xz, or .zip will be automatically uncompressed.
files starting with http://, ftp://, or ftps:// will be automatically
downloaded and decompressed.}
}
\value{
A data frame with two columns.The \code{id} column contains
the name for each sequence and the \code{taxonomy} column is a series of
taxonomic names separated by semi-colon.The string does not have a
semi-colon at the end of the sequence.
}
\description{
Read a \href{https://mothur.org/wiki/constaxonomy_file/}{mothur-formatted taxonomy file}
into as a data frame
}
\note{
There are no checks to insure that each sequence has a unique
id value. it is assumed that each sequence has the same number
of taxonomic levelsrepresented in the second column of the input file.
}
\examples{
temp <- tempfile()
write("seqA\tA;B;C;",file = temp)
write("seqB\tA;B; C;",file = temp, append = TRUE)
write("seqC\tA; B;C;",file = temp, append = TRUE)
write("seqD\tA;B;C",file = temp, append = TRUE)
write("seqE\tA;B; C",file = temp, append = TRUE)
write("seqF\tA; B;C",file = temp, append = TRUE)
write("seq G\tA;B;C;",file = temp, append = TRUE)
read_taxonomy(temp)

}
