% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{trainset19_db}
\alias{trainset19_db}
\title{trainset19_db}
\format{
A bacteriatypR database with 24,642 sequences that has been trained
using 8-mer kmers for the RDP taxonomy from phylum down to genus:
\describe{
\item{conditional_prob}{the conditional probabilities for each genus and kmer}
\item{genera}{taxonomic string with each level separated with a \verb{;}}
}
}
\source{
\itemize{
\item \href{https://mothur.org/rdp/wiki/rdp_reference_files/}{mothur_formatted files}
\item \href{https://mothur.org/blog/2024/RDP-v19-reference-files/}{Description of how mothur-formatted files were generated}
\item \href{https://sourceforge.net/p/rdp-classifier/news/2023/08/rdp-classifier-214-august-2023-released/}{RDP sourceforge page}
}
}
\usage{
trainset19_db
}
\description{
The sequence and taxonomy data for the 24,642 sequences found in the Ribosomal
Database Project's trainset19_072023 training set for use with the naive
bayesian classifier as implemented in the \code{{bacteriatypR}} R package.
generated using the \code{bacteriatypR::build_kmer_database()} function:
}
\keyword{datasets}
