% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmers.R
\name{classify_sequence}
\alias{classify_sequence}
\title{Title}
\usage{
classify_sequence(
  unknown = unknown_sequence,
  database = db,
  kmer_size = 8,
  num_bootstraps = 100
)
}
\arguments{
\item{kmer_size}{}
}
\description{
Title
}
